chore: export getTotals helper from @redocly/core
